pyproject.toml
[tool.poetry]
name = "dandori"
version = "0.1.0"
description = ""
authors = ["Hiroyuki Tanaka <aflc0x@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
ghapi = "^0.1.17"

[tool.poetry.dev-dependencies]
ipython = "^7.23.1"

[tool.poetry.scripts]
dandori = 'dandori.__main__:main'

[tool.pylint.SPELLING]
max-spelling-suggestions = 0

[tool.pylint.'MESSAGES CONTROL']
disable = ["R", "C0330", "C0412", "C0111", "arguments-differ", "fixme", "abstract-method"]

[tool.pylint.FORMAT]
max-line-length = 120

[tool.pylint.BASIC]
variable-rgx = "[a-z_][a-z0-9_]*$"
argument-rgx = "[a-z_][a-z0-9_]*$"
attr-rgx = "[a-z_][a-z0-9_]*$"

[tool.black]
line-length = 120
target-version = ["py39"]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
